+++ src/components/RequestLimitComponents/RequestLimit.tsx
@@ -1,9 +1,15 @@
-import React from 'react';
+/* eslint-disable no-unused-vars */
+import React from "react";
 
 import { useRequestLimitData } from "../../hooks/useRequestLimitData.ts";
 import { prepareRequestLimitChartData } from "../../utils/requestLimitUtils.ts";
-import RequestLimitSorter from "./RequestLimitSorter.jsx";
-import RequestLimitChart from "./RequestLimitChart.jsx";
+import RequestLimitSorter from "./RequestLimitSorter.tsx";
+
+import { getRequestLimitChartOptions } from "../../config/requestLimitChartConfig.ts";
+
+import mainStore from "../../stores/mainStore.ts";
+import dataStore from "../../stores/dataStore.ts";
+
 import {
   Chart as ChartJS,
   CategoryScale,
@@ -15,6 +21,8 @@ import {
   Legend,
 } from "chart.js";
 
+import { Bar } from "react-chartjs-2";
+
 ChartJS.register(
   CategoryScale,
   LinearScale,
@@ -25,14 +33,10 @@ ChartJS.register(
   Legend,
 );
 
-import mainStore from "../../stores/mainStore.ts";
-import dataStore from "../../stores/dataStore.ts";
-
 const RequestLimit = () => {
-
   const { selectedMetric } = mainStore();
   const requestLimits = dataStore((state) => state.allData.requestLimits);
-  const { sortType, setSortType } = dataStore();
+  const sortType = dataStore((state) => state.sortType);
 
   const podList = useRequestLimitData(requestLimits, sortType, selectedMetric);
   const [limitData, requestData] = prepareRequestLimitChartData(
@@ -58,16 +62,23 @@ const RequestLimit = () => {
     ],
   };
 
+  const chartHeight = podList.length * 50;
+  const options = getRequestLimitChartOptions(selectedMetric);
+
   return (
     <div className="flex flex-col gap-4">
       <div className="px-4">
-        <RequestLimitSorter setSortType={setSortType} />
+        <RequestLimitSorter />
       </div>
-      <RequestLimitChart
-        data={chartData}
-        selectedMetric={selectedMetric}
-        podList={podList}
-      />
+      <div className="overflow-y-auto p-4" style={{ height: chartHeight }}>
+            {podList.length > 0 ? (
+              <Bar options={options} data={chartData} />
+            ) : (
+              <div className="flex h-full items-center justify-center">
+                <p className="text-slate-900">No data available</p>
+              </div>
+            )}
+          </div>
     </div>
   );
 };
